envtest refactoring
