+++ .github/workflows/go.yml
@@ -91,7 +91,7 @@ jobs:
           path: install.yaml
 
   unittest:
-    name: Unit & Integration tests
+    name: Unit tests
     needs: build
     runs-on: ubuntu-latest
 
@@ -107,6 +107,24 @@ jobs:
       - name: Run unit tests
         run: |
           make test
+
+  envtest:
+    name: Integration tests
+    needs: build
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out code into the Go module directory
+        uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version-file: go.mod
+
+      - name: Run env tests
+        run: |
+          make envtest
 
   generate-sbom:
     name: "Build :: SBOM"
