+++ internal/controller/bootstrap/common_test.go
@@ -1,3 +1,5 @@
+//go:build envtest
+
 package bootstrap
 
 import (
