+++ internal/controller/bootstrap/suite_test.go
@@ -1,3 +1,5 @@
+//go:build envtest
+
 /*
 Copyright 2025.
