+++ internal/controller/controlplane/k0smotron_controlplane_controller_test.go
@@ -1,3 +1,5 @@
+//go:build !envtest
+
 package controlplane
 
 import (
