+++ internal/controller/k0smotron.io/k0smotroncluster_configmap_test.go
@@ -1,3 +1,5 @@
+//go:build !envtest
+
 /*
 Copyright 2023.
