+++ internal/controller/k0smotron.io/k0smotroncluster_entrypoint_test.go
@@ -1,3 +1,5 @@
+//go:build !envtest
+
 /*
 Copyright 2023.
