+++ internal/controller/k0smotron.io/k0smotroncluster_etcd_test.go
@@ -1,3 +1,5 @@
+//go:build !envtest
+
 /*
 Copyright 2023.
