+++ internal/controller/k0smotron.io/k0smotroncluster_kubeconfig_test.go
@@ -1,3 +1,5 @@
+//go:build !envtest
+
 package k0smotronio
 
 import (
